fix(templates): adds dependabot config and remove node makefile
Adds a dependabot configuration file to automate dependency updates for TypeScript, Python, and Go templates.

Updates the documentation links for Suga CLI installation in the Go and Python templates.

Upgrades the base Node.js version in the TypeScript Express template's Dockerfile to version 22.

Removes the Makefile from the TypeScript Express template and updates the README to use npm scripts for build tasks, providing instructions for yarn, pnpm, and bun.